// generated/src/app/api-models/cieb-currency-code.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebCurrencyCode {
  djCountryDesc: string;
  defaultTransPaymentCode: string;
  updateProcess: string;
  updateUser: string;
  updateDatetime: Date;
  insertProcess: string;
  insertUser: string;
  insertDatetime: Date;
  activeInd: string;
  incomingOnlyInd: string;
  wholeCurrencyInd: string;
  euroInd: string;
  currencyCodeDesc: string;
  currencyCode: string;
}